requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
